Renamed file helperStartList -> parseNames
